build 64-bit tar

+++ usr/src/cmd/tar/tar.c
@@ -8270,8 +8270,7 @@ read_xattr_hdr(attr_data_t **attrinfo)
 	    sizeof (struct xattr_hdr));
 	(void) sscanf(xattrp->h_namesz, "%7d", &namelen);
 	if (link_len > 0)
-		xattr_linkp = (struct xattr_buf *)
-		    ((int)xattrp + (int)comp_len);
+		xattr_linkp = (struct xattr_buf *)(xattrp + comp_len);
 	else
 		xattr_linkp = NULL;
